requirements.txt
streamlit==1.42.2
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.13.3
openai==1.64.0
python-dotenv==1.0.1
langchain==0.3.19
langgraph==0.3.2
pymongo==4.11.1
python-dateutil==2.8.2
feedparser==6.0.11
langchain-openai==0.3.7
boto3==1.34.28
awscli==1.32.28